bounds, total_bounds, estimate_utm_crs
